Rename scaffold-to-generate
