+++ docs/eslint-rules.md
@@ -10,11 +10,11 @@ This rule ensures that command flags have the environment variable set.
 This way users can decide between passing flags using arguments or environment variables:
 
 ```bash
-shopify app scaffold extension --type product_subscription
+shopify app generate extension --type product_subscription
 
 # vs
 
-SHOPIFY_FLAG_EXTENSION_TYPE=product_subscription shopify app scaffold extension
+SHOPIFY_FLAG_EXTENSION_TYPE=product_subscription shopify app generate extension
 ```
 
 ### `command-conventional-flag-env`
