+++ packages/app/src/cli/prompts/generate/extension.test.ts
