+++ packages/app/src/cli/services/scaffold/extension.ts
@@ -102,9 +102,9 @@ async function uiExtensionInit({
         },
       },
       {
-        title: `Scaffold ${getExtensionOutputConfig(extensionType).humanKey} extension`,
+        title: `Generate ${getExtensionOutputConfig(extensionType).humanKey} extension`,
         task: async (_, task) => {
-          task.title = `Scaffolding ${getExtensionOutputConfig(extensionType).humanKey} extension...`
+          task.title = `Generating ${getExtensionOutputConfig(extensionType).humanKey} extension...`
 
           const templateDirectory = await path.findUp(
             `templates/ui-extensions/projects/${mapExtensionTypeToExternalExtensionType(extensionType)}`,
@@ -128,7 +128,7 @@ async function uiExtensionInit({
             await changeIndexFileExtension(extensionDirectory, extensionFlavor)
           }
 
-          task.title = `${getExtensionOutputConfig(extensionType).humanKey} extension scaffolded`
+          task.title = `${getExtensionOutputConfig(extensionType).humanKey} extension generated`
         },
       },
     ],
@@ -181,7 +181,7 @@ async function functionExtensionInit(options: FunctionExtensionInitOptions) {
     const templateDownloadDir = path.join(tmpDir, 'download')
 
     await ui.task({
-      title: `Scaffolding ${getExtensionOutputConfig(options.extensionType).humanKey} extension...`,
+      title: `Generating ${getExtensionOutputConfig(options.extensionType).humanKey} extension...`,
       task: async () => {
         await file.mkdir(templateDownloadDir)
         await git.downloadRepository({
@@ -196,7 +196,7 @@ async function functionExtensionInit(options: FunctionExtensionInitOptions) {
           await file.remove(configYamlPath)
         }
         return {
-          successMessage: `${getExtensionOutputConfig(options.extensionType).humanKey} extension scaffolded`,
+          successMessage: `${getExtensionOutputConfig(options.extensionType).humanKey} extension generated`,
         }
       },
     })
